kvutils: Move FingerprintedReadSet.
