+++ ledger/participant-state/kvutils/src/main/scala/com/daml/ledger/participant/state/kvutils/package.scala
@@ -36,7 +36,6 @@ package object kvutils {
   type CorrelationId = String
 
   type Fingerprint = Bytes
-  type DamlStateMapWithFingerprints = Map[DamlStateKey, (Option[DamlStateValue], Fingerprint)]
   val FingerprintPlaceholder: Fingerprint = ByteString.EMPTY
 
   val MetricPrefix: MetricName = MetricName.DAML :+ "kvutils"
